Generally promote compile warnings to errors only on our 'development' version of gcc which is gcc 4.4.  Other versions of gcc should not get errors promoted.  (morlovich)
Re-interpret StringPieceToVector's delims arg as a set, rather than a
sequence, adding testcases to support that.  All our current usages were
single-character so this change has no current effect. (fangfei)

+++ src/net/instaweb/rewriter/image_rewrite_filter.cc
@@ -326,7 +326,8 @@ bool ImageRewriteFilter::CanInline(
     int image_inline_max_bytes, const StringPiece& contents,
     const ContentType* content_type, GoogleString* data_url) {
   bool ok = false;
-  if (content_type != NULL && contents.size() <= image_inline_max_bytes) {
+  if (content_type != NULL &&
+      static_cast<int>(contents.size()) <= image_inline_max_bytes) {
     DataUrl(*content_type, BASE64, contents, data_url);
     ok = true;
   }
